refactor to support generated docs
